+++ ufp/descriptor/__init__.py
@@ -4,3 +4,4 @@
 from __future__ import absolute_import
 
 from .class_property import classproperty, ClassPropertyMetaClass
+from .cached_property import CachedProperty